pack for macos and electrun-builder two package structure

+++ scripts/clean.ts
@@ -1,5 +1,6 @@
 import * as fs from "fs";
 import { join, resolve } from "path";
+import { outDir, distDir } from "./common";
 
 async function exists(path: fs.PathLike): Promise<boolean> {
   try {
@@ -10,18 +11,29 @@ async function exists(path: fs.PathLike): Promise<boolean> {
   }
 }
 
-async function rmRecursively(path: string) {
+async function rmRecursively(path: string, excludes?: Array<string>) {
   if (!(await exists(path))) {
     return;
   }
   if (path.substr(0, 1) !== "/" && path.indexOf(":") === -1) {
     path = resolve(path);
   }
+  const excludeFiles = (excludes ?? []).map((i) => {
+    return resolve(path, i);
+  });
   let files: Array<string> = [path];
   while (files.length > 0) {
     const last = files.pop();
-    const stat = await fs.promises.stat(last);
-    if (!stat.isDirectory()) {
+    const stat = await fs.promises.lstat(last);
+    if (last === path && !stat.isDirectory) {
+      return;
+    }
+    if (excludeFiles.includes(last)) {
+      continue;
+    }
+    if (stat.isSymbolicLink()) {
+      await fs.promises.unlink(last);
+    } else if (!stat.isDirectory()) {
       await fs.promises.rm(last);
     } else {
       const children = await (await fs.promises.readdir(last)).map((p) => {
@@ -30,11 +42,17 @@ async function rmRecursively(path: string) {
       if (children.length === 0) {
         await fs.promises.rmdir(last);
       } else {
-        files.push(last);
+        if (last !== path) {
+          files.push(last);
+        }
         files = files.concat(children);
       }
     }
   }
+  if ((await fs.promises.readdir(path)).length === 0) {
+    await fs.promises.rmdir(path)
+  }
 }
 
-rmRecursively("./dist");
+rmRecursively(outDir, ["package.json", "yarn.lock", "package-lock.json"]);
+rmRecursively(distDir);
